requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Faker==8.14.0
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.3.1
pixelmatch==0.2.3
playwright==1.14.1
pluggy==0.13.1
py==1.10.0
pyee==8.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-base-url==1.4.2
pytest-ordering==0.6
pytest-playwright==0.2.0
python-dateutil==2.8.2
python-slugify==5.0.2
requests==2.25.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
websockets==9.1
